cmd/snap-confine: add more flexible namespace manager
This patch adds new namespace manager. Unlike the old code it can work
with namespaces with less boilerplate as it holds fd's to essential
firectory and lock file around.

The new manager can also enumerate and discard namespaces with greater
ease.It is also using error propagation so should be less tedious to
test.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>