cmd/snap-discard-ns: add support for --all
This patch changes snap-discard-ns to behave in a special way when
invoked with "--all" instead of a snap name. Since "--all" is not a
valid snap name this can be detected correctly.

When invoked with --all the namespace manager is queried to enumerate
all the possible namespaces (snap names). Each such namespace is
discarded, one by one.

When invoked without --all nothing is changed from pervious behavior.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>