cmd/libsnap: add sc_quote_string
This patch adds a function for safely quoting valid C strings that are
coming from unknown sources. The output is similar to python's repr()
where most typical printable characters are unmodified with the
following exceptions.

The double-quote character and various whitespace (e.g. newline)
characters are escaped with their typical escape sequences. The escape
character is also escaped the same way. Every other character is escaped
with the syntax "\xff" where "ff" is the hexadecimal code of the
character.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>