interfaces: move some policy tests to buildin package
TL;DR; this makes people that "go test" in builtin, while working on new
interfaces, get a more complete picture of what is working.

This patch starts the move of interface-speicfic tests that check base
policy into the builtin package. This is more natural and goes back to
the old idea of keeping all of the interface code in a pair of files
(interface and its tests). Theere are many other things to improve here.
The common tooling is duplicated and inadequate, some things should move
to snaptest package. The contents of the policy_test.go file should be
moved to individual interfaces and to the all_test.go file. This will
happen over time.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>