snap-update-ns: WIP on generalized mount profiles
This unfinished change lets snap-update-ns create mounts in nearly any
location. Writable holes are created automatically using overlayfs. They
are also undone when the corresponding entry goes away.

This will allow most of the new layout features to operate. It will also
allow to improve the content interface to implement the aggregation
features (through a .d-style directories).

There are lots of thins missing still:
 - security review needs to identify things we should refuse to
   support (such as bi-directionally shared mount locations)
 - certain things in overlayfs are not allowed but also not checked for
   (see kernel docs for detials). Those need to be implemented as checks
   here.

This patch adds special handling of overlays to the routine that
computes the set of needed mount changes. The overlays are synthesized
by snap-update-ns and are not present in desired profiles but now they
will be preserved whenever snap-update-ns has at least one mount entry
(typically a bind mount) underneath the overlay.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>