tests: fix unmounting on systems without rshared /
This patch adds a test that reproduces a bug in snap-confine/snapd that
affects Ubuntu 16.04 running an LXD container with snapd. Inside such
container the root directory is not implicitly rshared by systemd and
after running any snap command snaps cannot be removed/unmounted
correctly.

The error is addressed by a little trick in snap-confine. The existing
approach that bind-mounts /snap over itself and then rshares it is
replaced with a loop that rshares each mounted snap. In effect /snap
bind mount doesn't exist anymore and all application and hook processes
start with the sharing already in place.

With-Kind-Regards-To: Kyle Fazzari
Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>