snap-update-ns: poke writable holes with tmpfs
This unfinished change lets snap-update-ns create mounts in nearly any
location. Writable holes are created automatically using a "writable
mimic" which is explained below. They are also undone when the
corresponding entry goes away.

A "writable mimic" is a bind-mount farm created on top of tmpfs that
hides the original, read-only directory. This is a simple and portable
solution in absence of correctly operating overlayfs.

This will allow most of the new layout features to operate. It will also
allow to improve the content interface to implement the aggregation
features (through a .d-style directories).

TODOs:
 - undo works but needs changes to do path to inject
   x-snapd.{id,parent-id,synthetic} options to let us track things.
 - there is no undo for symlinks yet
 - apparmor rules need a review
 - spread test needs adjustments to have symlink and file

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>