many: WIP on conservative mount
This WIP patch represents a snapshot on conservative mount, a feature
where mount call is preceded by some checks and potential races are
discovered after-the-fact.

The patch is huge because I chose to add the feature to osutil where it
can be of general use. I will work on splitting the move of the test
machinery from cmd/snap-update-ns to osutil.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>