cmd/snap-confine: allow hard-coded mounts to be optional
This patch changes the code path processing the list of built-in mount
operations so that certain operations are allowed to fail if the mount
point or mount source are missing.

This is convenient when making base snaps as we now have a vocabulary to
say that something is actually optional.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>