overlord/ifacestate: remove implicit slots
This patch removes the support infrastructure that added implicit slots
throughout the code. The idea is that with the core18 transition looming
we will create all formerly implicit slots on a new virtual "system"
snap (which doesn't really exist and is restricted from existing).

Existing state is unaffected, it will still hold connections that
reference the "core" snap. The connection reloading logic is adjusted to
translate that to "system".

The new system snap is virtually added to the repository alone. It
doesn't exist in the state. The few sites that look up snap info for the
purpose of setting up security skip the operation when invoked for the
"system" side. This is safe since, again, by definition this virtual
snap doesn't exist and doesn't have any security that needs to be setup
(no applications, no profiles).

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>