data/completion: fix inconsistency in +x and shebang (#5353)
* data/completion: remove shebang and +x from etelpmoc.sh

The $(rev<<complete.sh) script doesn't need to be executable as it is
invoked directly via bash. It also doesn't need a shebang line for the
very same reason.

This silences the following rpmlint warning:

[  108s] snapd.x86_64: W: non-executable-script /usr/lib/snapd/etelpmoc.sh 644 /bin/bash
[  108s] This text file contains a shebang or is located in a path dedicated for
[  108s] executables, but lacks the executable bits and cannot thus be executed.  If
[  108s] the file is meant to be an executable script, add the executable bits,
[  108s] otherwise remove the shebang or move the file elsewhere.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>

* data/complete,packaging: make complete scripts +x, with shebang

This undoes earlier patch but this is the best compromise I managed to
find. With this approach the scripts are executable (even though we
don't rely on that) and have a valid shebang. This makes both rpmlint
and shellcheck happy.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>

* data/complete: match +x in the packaging with the tree

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>

* data/complete,packaging: make complete scripts -x, sans shebang

The shebang conundrum was finally solved with a shellcheck specific
markup that identifies the shell type. This allows the scripts to be
shellcheck-clean, not executable and devoid of shebang. This allows
rpmlint to become happy again.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>