cmd/snap-confine: fix snaps running on core18
This patch changes the logic that selects mount namespace construction
behavior so that it is correct on machines booting with core18 wishing
to run core16/core snaps.

In the past core devices never used pivot_root and always used a minimal
set of changes on top of the initial mount namespace. Once bases were
introduced that logic was extended to use a full pivot_root approach
(which is always used on classic) when the invoked snap uses base other
than "core" (since the boot base was fixed at the time).

With the introduction of bootable core18 base, snaps need to use
pivot_root and the regular set of bind mounts to see their expected
"core" root filesystem when running on core18.

The code is tweaked to introduce "normal mode" and "legacy mode". Normal
mode uses pivot root and a predictable set of bind mounts. It is always
used on classic, it is also used on core devices _except_ for one
exception that limits any chances of backwards incompatible changes to
existing core devices.

A core device using "core" as boot base running snaps that demand "core"
base snap will use the "legacy mode" which doesn't pivot root. This is
exactly what used to happen on core devices before.

Core devices are identified with os-release(2) ID=ubuntu-core string.
A "core16" or "core" device is identified with an additional
VERSION_ID=16 string. All devices without ID=ubuntu-core are considered
classic. This logic will likely need to be extended as the Fedora 29
base work advances but no decision on how to identify such devices has
been made yet.

Existing classification between core and classic is extended to
differentiate core16 from other core systems. Classic classification is
retained as-is. There is a new helper function for checking if normal
mode should be used, given distribution and base snap name. Existing
call sites that used classic distinction are changed to use it instead.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>