many: remove build-time choice of SNAP_MOUNT_DIR
This patch removes one build-time distinction from the C parts of snapd,
namely the distinction of where snaps are mounted.

In the ancient past all snaps would always be mounted at /snap and all
the code was simple and correct. Then came the dragon of complexity and
added a variant.  Some distributions kept using /snap while others
started to use /var/lib/snapd/snap and the world became fractured. In
our wisdom we added a build-time option and used sed and other dark
magic to replace the custom variable and make the package compliant and
operational. This also broke the ability to use snaps to update snapd
itself for those people.

Then came the knights of enlightenment and decided to kill the dragon by
with the blade of convention and sanity. From now on instead of
compile-time open choices snapd would have built-in support for a small
set of options so that one build could be universally used everywhere.
People cheered and there was much rejoicing.

The knights of enlightenment now set their eye on the Nvidia support
code, /media handling and other monsters that split people into tribes
of incompatibility.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>