cmd/snap-update-ns: switch to user requested by -u
When snap-update-ns is invoked with -u it is an indication that we are
going to run as the specified user ID. This also implies --user-fstab
and lack of --from-snap-confine. In result we are going to change to the
specified user ID, join the appropriate mount namespace and perform the
updates in the context of that user.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>