cmd/snap-update-ns: manually implement issspace (!!!)
I wasn't able to find any other cause but using the isspace macro,
outside of tests (wich pass just fine), causes ./snap-update-ns -u 1000
to crash. The debugger points to *uid_text being the culprit despite
it not being a NULL pointer. It is correctly pointing to the string
"1000" when the crash happens. I was unable to solve it with a static
function that calls isspace internally (which is probably a macro).

As such, implement isspace ourselves.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>