cmd/snap-confine: preserve/restore uid/gid centrally.
Currently snap-confine is a setuid root and a setgid root executable.
Oddly so, it also doesn't really use this property, choosing to drop the
group permission on startup.

This patch moves the UID/GID handling to the process state handling. On
startup we remember the set of UID/GID values and switch effective GID
to root. On restore we carefully restore the correct values.

This allows snap-confine to stop being setgid root. This is reflected in
packaging, where appropriate.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>
Related-To: https://bugzilla.suse.com/show_bug.cgi?id=1127368