snap: fix interface bindings on implicit hooks
An implicit hook is one that exists in the filesystem but is not
mentioned in the meta/snap.yaml file.

An interface binding is an association of a plug or slot definition with
an app or hook.

Plugs and slots can be defined either at the top level using the full
syntax (comprised of key-value pairs) or at an app or hook level using
the abbreviated syntax (with the interface name alone).

When an application or hook references a top-level plug or slot, or
defines a new plug or slot named after the interface name, that act
automatically scopes the interface to a particular application or hook.

This feature allows snap developers to implement a form of privilege
separation, where particular applications or hooks have a subset of the
permissions resulting from the set of interface connections.

In theory implicit hook definitions are no different from explicit hook
definitions. Due to a bug however an implicitly defined hook would get
*all* the top-level plugs and slots defined in a given snap, even if
those interfaces were already scoped to some applications. This would
not happen if said hook was explicitly defined in the YAML and not
bound to any plug or slot.

This patch fixes this by not considering top-level plugs and slots that
already have app or hook associations. Such interfaces must be already
scoped to a particular app or hook so they would not be assigned to an
hook if it was explicitly defined.

Fixes: https://bugs.launchpad.net/snapd/+bug/1824557
Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>