cmd/snap-update-ns: switch to traditional Lock/Unlock
In the past the Lock function would return the unlock function and an
error, if any. This was regarded as odd by Chipaca during code review.
I promised a follow-up and here it is, the flock-based lock is held
internally by the mount profile update context. Lock no longer returns
the unlock function. Unlock is now a part of the update context
interface.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>