many: add no-legacy-nvidia-hack feature
LOTS OF STUFF NOT RELATED TO ^^^ ADDED

This new, exported feature, when enabled, disables all of the special
nvidia code that allows reusing userspace parts of the host nvidia driver.

This part of snap-confine is tricky and complex and is a source of
constant headache. As we are moving towards general GPU support driven
by snapd, it is also slowly becoming unnecessary.

Enabling the feature simply prevents snap-confine from either
bind-mounting /usr/lib/nvidia-NNN or constructing the symlink farm from
various host libraries.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>