httputil: improve detection of unavailable DNS servers
When we detect if the DNS server is unavailable, in response to a HTTP
request error, we unpack the returned error object, hoping to reach
net.DNSError. When the error message contained therein contains the text

    Temporary failure in name resolution

We understand that this really conveys the fact we cannot talk to the
DNS sever. This error message, however, depends on the golang DNS
backend, either native or cgo. Similar check in ShouldRetryError
carries a separate test for another message:

    connection refused

When running unit tests inside the open build service the network setup
is such that we hit this case exactly, with the returned errors being:

    [   44s] nesting 2: &net.DNSError{Err:"read udp [::1]:58819->[::1]:53: read: connection refused", Name:"nonexistingserver909123.com", Server:"[::1]:53", IsTimeout:false, IsTemporary:true}

This patch extends the check to handle both situations.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>