many: use a dedicated named cgroup hierarchy for tracking
This patch changes how snap-confine and snapd perform process tracking
required for the refresh-app-awareness feature. Instead of abusing the
pids cgroup that systemd may also want to use, and stealing the process
from the cgroups managed by systemd, we now use the following approach.

We mount a new hierarchy /run/snapd/cgroup. As a side note, the earlier
review of this patch used /sys/fs/cgroup/snapd but this location poses
issues when used in pure-v2 mode, where it would seem like a
sub-hierarchy of cgroup2 file-system. All prior use of
/sys/fs/cgroup/pids/ is switched to use /run/snapd/cgroup instead.

Various management scripts are updated to unmount this hierarchy when
snapd is purged from the system.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>