cmd/snapd-release-agent: add cgroup release agent
This patch adds the cgroup release agent.

A release agent is a cgrup v1 concept. When a cgroup v1 hierarchy
becomes empty, housing no further hierarchies or processes, and when the
release notification is enabled then the hierarchy-wide release agent is
invoked by the kernel with the name of the hierarchy that became empty.

This mechanic is used to solve a number of long-standing issues:

We can now remove unused cgroup hierarchies, namely the device, freezer,
and the recently introduced name=snapd one. This releases memory and can
be of benefit for long running systems that install and remove large
number of snaps

We can now remove apparmor profiles from the kernel. This is another
issue of resource management. Note that unlike cgroups, which are
created and removed on demand, apparmor profiles are only unloaded if
the snap was already removed from the system.

In addition this mechanic allows us to get real-time notification of
application startup and termination. Please refer to the patch for the
detailed description but the main idea is that snap-confine and
snapd-release-agent respectively create and remove monitor files stored
in /run/snapd/monitor. This allows any 3rd party to use inotify to
observe changes. Specifically this allows snapd to implement live retry
of pending refresh as soon as an application is closed.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>