cmd: embed snapd-release-agent into "snap"
This way we get free re-execution support and one less binary. On the
downside the binary is *huge* and needs special handling as the kernel
cannot execute the release agent as a sub-command of another binary.
Therfore a symlink trick is used, where a symlink called
snapd-release-agent is used to trigger the new logic.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>