cmd: implement snap run --explain
This patch adds a new command line option for "snap run", the new
"--explain" toggle propagates through snap run, snap-confine and
snap-exec. At each stage it shows additional output that explains what
is going on. The elements that are highlighted are of importance to the
execution environment of the application process.

The purpose of this is to educate application developers about,
possibly unfamiliar, things that happen during application startup.

Forum: https://forum.snapcraft.io/t/snap-run-explain/13427
Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>