tests: port user-mounts test to session-tool
This test was failing for me recently so I decided to take a look and
*oh my*, so much stuff that can be simplified:

 - hand-coded "run as user" code replaced with session-tool
 - snap-discard-ns invoked with snap-tool
 - test snap installed with install_local rather than with snap try

In addition I made some stylistic changes, assertions that use "[ ... ]"
were changed to "test ..." as that form is more natural outside of an if
statement. The helper script that was hand-crafted is gone and the four
lines it invoked now run via session-tool, so that on failure we could
see exactly which line failed and why.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>