tests: port interfaces-desktop-* to session-tool
Those test were leaking the dbus-daemon process as they were running as
root and caused the root user to socket-activate dbus-daemon responsible
for the session-bus.

The usage of "session-tool -u test" ensures we have a session bus and
more importantly that it gets shut down at the end of the test.

This was detected with invariant-tool.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>