tests: detect tests that leave systemd in "needs-reload" state
Editing systemd units may leave it a state where unit definition in memory
does not match unit definition on disk. This can cause failures,
as subsequent tests will operate on the in-memory "ghost" rather than
what is on disk.

This can be detected by enumerating all units and checking the
NeedDaemonReload unit property.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>