tests: expand mount-ns test to cover ubuntu-core-18
The test relies on a set of data files that contain expected contents
of the mount table, transformed for determinism and observed from the
point of view of a per snap and per snap-and-user mount namespace of
a core16-based snap, core18-based snap and a classic snap.

The set of existing files was out of date, as the test was disabled for
a long while. It is now up-to-date, and the name reflects the trend to
use distinct backends to spread the overall load across distinct areas
of google cloud.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>