spread: do not install locally built snapd by default
XXX: This should be scoped to systems where we can re-execute.

The aim of this change is to allow testing the more realistic case,
where snapd that came with the distribution is used as a trampoline to
execute snapd that came from the snapd snap. The old way of testing can
easily mask mistakes that only manifest when people actually use the new
snapd, such as when introducing a change to an area that does not
participate in re-execution, such as a new service, the udev area or
similar.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>