cmd/snap-update-ns: add test explaining content-layout bug
Snapd is affected by a bug related to content-layout interaction.
There's a very popular scheme where a snap has two interacting elements
that only work in one special edge case and break regularly:

- A content plug is used to receive content of another snap, typically
  under $SNAP/foo somewhere. Where foo exists and is not created with a
  mimic.

- A layout entry is used to "distribute" $SNAP/foo to /usr/share/foo or
  other locations.

This particular arrangement is used to share essential libraries from a
common "runtime" snap which is not a dedicated base. The error arises in
one of several cases:

- The content is initially connected. Then disconnected and finally
  re-connected. There are two bugs here (references are in the code).

- TODO: The content is initially disconnected, a hook or the user
  initialized the namespace. The content is then connected leading to
  non-working snap.

- TODO: The content is initially connected and the revision of the
  content far side changes.

- TODO: The content is initially connected and the revision of the
  application changes.

There are several reasons why this breaks. The new tests outline this.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>