cmd/snap-update-ns: keep order of existing mount entries
We recently realized that the order of saved mount profile, after
performing mount namespace update, was subtly incorrect. The order was
seemingly swapped after each operation. In reality, the order reflected
the order of performed changes, including the special change type of
"keep", which means that an entry is kept intact.

Fix the problem by storing "kept" changes in the original order, which
is back-to-front compared to the order of changes computed by the
algorithm. At the same time, keep track of mount changes in the order of
their execution.

This probably fixes a hell lot of edge cases and weird behaviors that
went undiagnosed for many years.

Jira: https://warthogs.atlassian.net/browse/SNAPDENG-31644

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>