cmd/snap-confine, interfaces: re-share /snap
When snap-confine sets up the mount namespace, it constructs the /snap
directory as a view of the host's /var/lib/snapd/snap or /snap (a recursive
bind mount) in a manner where mount events are propagated from the parent mount
namespace (from the host) into the per-snap mount namespace, but not back.

This is logical as it allows snaps to see new revisions being mounted or old
revisions being unmounted without also polluting the host with any content
mounts that are often attached to $SNAP.

Before this patch, the /snap directory, as seen from the snap mount namespace
was not re-shared though. In consequence if a bind mount is taken from
somewhere in the $SNAP tree - like layouts often are - and then a content is
mounted in $SNAP then such mount event would not propagate to the layout.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>