many: allow bidirectional mounts from /run/user
Core desktop is exploring the ability to run some of the portal logic from a
snap that is not running as an unconfined hack. One of the properties of the
snap is that it must mount things in /run/user/*/ in a way that is visible to
other snaps. Snapd has had the ability do offer bidirectional mounts on a
per-mount basis but the set was fairly small, mostly /run/media or /media and
later on /run/netns.

Introduce an experimental feature flag 'bidirectional-mount-run-user' and teach
snap-confine to respect it.

This patch comes with no integration tests and is meant for early preview.

Jira: https://warthogs.atlassian.net/browse/SNAPDENG-34579

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>