snap: exclude security.selinux xattrs when building snapd
This allows building snapd on a SELinux system like Fedora and still
installing the snap on a non-SELinux system without hitting an error
such as this oner

```
error: cannot read snap file: cannot run unsquashfs: ----- FATAL ERROR: write_xattr: failed to
       write xattr security.selinux for file /tmp/read-file136240501/unpack/meta/snap.yaml because
       Operation not permitted Parallel unsquashfs: Using 18 processors 1 inodes (1 blocks) to
```

This patch requires updated squashfs-tools that ship in the snapd snap
to work correctly. It is not mergable as-is.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>