store: add support for forever cache
When the environment variable SNAPD_FOREVER_CACHE_DIR is not empty, then
the path designates a cache directory that supplements the existing
cache mechanism with a forever-cache. When used in CI/CD environment,
this allows installing snaps, purging snapd state and re-installing
snaps without incurring extra cost and unreliability of network
operations.

Improve the behavior of nullCache to avoid having to use pointer
instance.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>