osutil/vfs: refactor how mounts are attached
One of the realizations from last weeks experiments were that mounts
need to be explicitly attached to the parent mount but the logic to look
up the parent, so far using the path dominator logic, is not always the
same.

Refactor attachMount that is shared by both Mount and BindMount, with
the right signature to work with upcoming propagation support.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>